+++ lume/model.py
@@ -1,6 +1,5 @@
 from abc import ABC, abstractmethod
 from typing import Any
-from beamphysics import ParticleGroup
 
 from lume.variables import Variable
 
@@ -176,30 +175,3 @@ class LUMEModel(ABC):
         )
 
 
-class InitialParticlesMixIn(ABC):
-    """
-    Mix in to LUMEModel to indicate support for initial particles. Implement the following
-    property setter and getter in the concrete class.
-
-    Note: changing the initial particle distribution should update the models state.
-    """
-    @property
-    @abstractmethod
-    def initial_particles(self) -> ParticleGroup:
-        ...
-
-    @initial_particles.setter
-    @abstractmethod
-    def initial_particles(self, val: ParticleGroup):
-        ...
-
-
-class FinalParticlesMixIn(ABC):
-    """
-    Mix in to LUMEModel to indicate support for final particles. Implement the following
-    property in the concrete class.
-    """
-    @property
-    @abstractmethod
-    def final_particles(self) -> ParticleGroup:
-        ...
