+++ lume/staged_model.py
@@ -1,7 +1,35 @@
+from abc import ABC, abstractmethod
 from typing import Any
 
-from lume.model import FinalParticlesMixIn, InitialParticlesMixIn, LUMEModel
+from beamphysics import ParticleGroup
+
+from lume.model import LUMEModel
 from lume.variables.variable import Variable
+
+
+class InitialParticlesMixIn(ABC):
+    """
+    Mix in to LUMEModel to indicate support for initial particles.
+    """
+    @property
+    @abstractmethod
+    def initial_particles(self) -> ParticleGroup:
+        ...
+
+    @initial_particles.setter
+    @abstractmethod
+    def initial_particles(self, val: ParticleGroup):
+        ...
+
+
+class FinalParticlesMixIn(ABC):
+    """
+    Mix in to LUMEModel to indicate support for final particles.
+    """
+    @property
+    @abstractmethod
+    def final_particles(self) -> ParticleGroup:
+        ...
 
 
 class StagedModel(LUMEModel):
